Extract method to util package
