+++ internal/github/pullrequest/pull_request.go
@@ -3,10 +3,9 @@ package pullrequest
 import (
 	"context"
 
+	"github.com/agilepathway/label-checker/internal/util"
 	"github.com/shurcooL/githubv4"
 	"golang.org/x/oauth2"
-
-	"github.com/agilepathway/label-checker/internal/error"
 )
 
 type pullRequest struct {
@@ -43,7 +42,7 @@ func (pr pullRequest) labels() []string {
 	}
 
 	err := pr.apiClient.Query(context.Background(), &query, variables)
-	error.PanicIfError(err)
+	util.PanicIfError(err)
 
 	labelNodes := query.Repository.PullRequest.Labels.Nodes
